reafctor: bigdecimal ao inves de double nas vendas

+++ src/main/java/school/sptech/crud_proj_v1/projection/DesempenhoFuncionarioProjection.java
@@ -1,8 +1,10 @@
 package school.sptech.crud_proj_v1.projection;
 
+import java.math.BigDecimal;
+
 public interface DesempenhoFuncionarioProjection {
     String getVendedor();
     Integer getTotalVendas();
-    Double getTotalFaturado();
-    Double getComissaoTotal();
+    BigDecimal getTotalFaturado();
+    BigDecimal getComissaoTotal();
 }
