+++ src/main/java/school/sptech/crud_proj_v1/projection/FaturamentoTempoProjection.java
@@ -1,6 +1,8 @@
 package school.sptech.crud_proj_v1.projection;
 
+import java.math.BigDecimal;
+
 public interface FaturamentoTempoProjection {
     String getPeriodo();
-    Double getFaturamento();
+    BigDecimal getFaturamento();
 }
