+++ src/main/java/school/sptech/crud_proj_v1/projection/GraficoTempoProjection.java
@@ -1,8 +1,10 @@
 package school.sptech.crud_proj_v1.projection;
 
+import java.math.BigDecimal;
+
 public interface GraficoTempoProjection {
     Integer getAno();
     Integer getMes();
     Integer getSemana();
-    Double getFaturamentoTotal();
+    BigDecimal getFaturamentoTotal();
 }
