+++ src/main/java/school/sptech/crud_proj_v1/projection/MargemCategoriaProjection.java
@@ -1,6 +1,8 @@
 package school.sptech.crud_proj_v1.projection;
 
+import java.math.BigDecimal;
+
 public interface MargemCategoriaProjection {
     String getCategoria();
-    Double getMargem();
+    BigDecimal getMargem();
 }
