+++ src/main/java/school/sptech/crud_proj_v1/projection/MetodoPagamentoProjection.java
@@ -1,7 +1,9 @@
 package school.sptech.crud_proj_v1.projection;
 
+import java.math.BigDecimal;
+
 public interface MetodoPagamentoProjection {
     String getMetodo();
     Integer getQtdVendas();
-    Double getValorTotal();
+    BigDecimal getValorTotal();
 }
